refactor(265): split checkout page into modules
